refactor: rename source files folder
