+++ ciphering-cli-tool/src/cipher/cipher-helpers.js
