+++ ciphering-cli-tool/src/cipher/cipher.js
