+++ ciphering-cli-tool/src/cipher/index.js
