+++ ciphering-cli-tool/src/cli/cli.js
