+++ ciphering-cli-tool/src/cli/index.js
