+++ ciphering-cli-tool/src/transform/helpers.js
