+++ ciphering-cli-tool/src/transform/index.js
