+++ ciphering-cli-tool/src/transform/rot-transformer.js
