+++ ciphering-cli-tool/src/validate/index.js
