+++ ciphering-cli-tool/src/validate/validate-errors.js
