+++ ciphering-cli-tool/src/validate/validate-helpers.js
